+++ email/org.wso2.carbon.transport.email/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>org.wso2.carbon.transport.email.parent</artifactId>
         <groupId>org.wso2.carbon.transport</groupId>
-        <version>4.4.13-SNAPSHOT</version>
+        <version>4.4.23-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
@@ -16,11 +16,6 @@
     <description>Carbon Email Transport</description>
 
     <dependencies>
-        <dependency>
-            <groupId>org.wso2.carbon.config</groupId>
-            <artifactId>org.wso2.carbon.config</artifactId>
-            <version>1.0.0</version>
-        </dependency>
         <dependency>
             <groupId>javax.mail</groupId>
             <artifactId>javax.mail-api</artifactId>
@@ -64,10 +59,6 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.orbit.org.yaml</groupId>
-            <artifactId>snakeyaml</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.thoughtworks.icegreen</groupId>
             <artifactId>greenmail</artifactId>
             <version>1.3</version>
@@ -92,6 +83,7 @@
         </carbon.component>
     </properties>
 
+<!--
     <build>
         <plugins>
             <plugin>
@@ -140,6 +132,7 @@
             </resource>
         </resources>
     </build>
+    -->
 
 
 </project>
