--- email/org.wso2.carbon.transport.email/src/main/java/org.wso2.carbon.transport.email/config/generator/Pop3ServerConfiguration.java
@@ -1,40 +1,0 @@
-package org.wso2.carbon.transport.email.config.generator;
-
-import org.wso2.carbon.config.annotation.Configuration;
-import org.wso2.carbon.config.annotation.Element;
-
-import java.util.Locale;
-
-/**
- * Created by chathurika on 7/20/17.
- */
-@Configuration(description = "POP3 Server Configurations")
-public class Pop3ServerConfiguration {
-    @Element(description = "name of the parameter")
-    private String name;
-
-    @Element(description = "value of the parameter")
-    private String value;
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    @Override
-    public String toString() {
-        return String.format(Locale.ENGLISH, "name : %s, value : %s",
-                name, value);
-    }
-}
