--- email/org.wso2.carbon.transport.email/src/main/java/org.wso2.carbon.transport.email/utils/EmailConstants.java
@@ -1,78 +1,0 @@
-package org.wso2.carbon.transport.email.utils;
-
-/**
- * This class contains the constants related to File transport.
- */
-public class EmailConstants {
-
-    /**
-     * Connector property
-     */
-    public static final String PROTOCOL_EMAIL = "email";
-    public static final String POLLING_INTERVAL = "pollingInterval";
-    public static final String SERVICE_NAME = "serviceName";
-    public static final String MAX_RETRY_COUNT = "maxRetryCount";
-    public static final String RETRY_INTERVAL = "reconnectIntervel";
-    public static final String CONTENT_TYPE = "contentType";
-
-    /**
-     * Email Receiver properties
-     */
-    public static final String MAIL_RECEIVER_USERNAME = "receiver.username";
-    public static final String MAIL_RECEIVER_PASSWORD = "receiver.password";
-    public static final String MAIL_RECEIVER_STORE_TYPE = "receiver.storeType";
-    public static final String MAIL_RECEIVER_HOST_NAME = "receiver.hostName";
-    public static final String MAIL_RECEIVER_FOLDER_NAME = "receiver.folderName";
-
-    /**
-     * Email Receiver properties
-     */
-    public static final String MAIL_SENDER_USERNAME = "sender.username";
-    public static final String MAIL_SENDER_PASSWORD = "sender.password";
-    public static final String MAIL_SENDER_HOST_NAME = "sender.hostName";
-
-   /* // Email message properties
-    public static final String FROM = "from";
-    public static final String TO = "to";
-    public static final String BCC = "bcc";
-    public static final String CC = "cc";
-    public static final String SEND_DATE = "sendDate";
-    public static final String SUBJECT = "subject";
-    //public static final String IS_EXPUNGED = "isExpunged";*/
-
-    /**
-     * Mail Headers which has to set in the message to be send
-     */
-    public static final String MAIL_HEADER_TO = "To";
-    public static final String MAIL_HEADER_FROM = "From";
-    public static final String MAIL_HEADER_CC = "Cc";
-    public static final String MAIL_HEADER_BCC = "Bcc";
-    public static final String MAIL_HEADER_REPLY_TO = "Reply-To";
-    public static final String MAIL_HEADER_IN_REPLY_TO = "In-Reply-To";
-    public static final String MAIL_HEADER_SUBJECT = "Subject";
-    public static final String MAIL_HEADER_MESSAGE_ID = "Message-ID";
-    public static final String MAIL_HEADER_REFERENCES = "References";
-    public static final String MAIL_HEADER_CONTENT_TYPE = "Content-Type";
-
-    /**
-     * Properties which are included in carbon message other than headers
-     */
-    public static final String MAIL_PROPERTY_FLAGS = "flags";
-    public static final String MAIL_PROPERTY_MESSAGE_NUMBER = "messageNumber";
-    public static final String MAIL_PROPERTY_UID = "messageUID";
-    public static final String MAIL_PROPERTY_REPLY_TO = "replyTo";
-    public static final String MAIL_PROPERTY_RECEIVED_DATE = "receivedDate";
-
-    public static final String ACTION_AFTER_PROCESSED = "actionAfterProcessed";
-    public static final String MOVE_TO_FOLDER = "moveToFolder";
-
-    public static final String AUTO_ACKNOWLEDGE = "autoAcknowledge";
-
-    /**
-     * Enum for action
-     */
-    public enum ActionAfterProcessed {
-        MOVE, SEEN, DELET, FLAGGED, ANSWERED, KEEPSAME
-    }
-
-}
