+++ email/org.wso2.carbon.transport.email/src/main/resources/META-INF/services/org.wso2.carbon.messaging.ClientConnector
@@ -1,0 +1,1 @@
+org.wso2.carbon.transport.email.sender.EmailClientConnector
