+++ email/org.wso2.carbon.transport.email/src/main/resources/META-INF/services/org.wso2.carbon.messaging.ServerConnectorProvider
@@ -1,0 +1,1 @@
+org.wso2.carbon.transport.email.provider.EmailServerConnectorProvider
