--- email/org.wso2.carbon.transport.email/src/test/java/org/wso2/carbon/transport/email/test/Utils/EmailTestUtils.java
@@ -1,27 +1,0 @@
-package org.wso2.carbon.transport.email.test.Utils;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Created by chathurika on 7/11/17.
- */
-public class EmailTestUtils {
-
-  /*  public Map<String, String> getEmailReceivedServerParameters(String username,
-            String password, String host, String store, String folder) {
-        Map<String, String> emailReceivedServerParameters = new HashMap<String, String>();
-        emailReceivedServerParameters.put(EmailTestConstant.USERNAME, username);
-        emailReceivedServerParameters.put(EmailTestConstant.PASSWORD, password);
-        emailReceivedServerParameters.put(EmailTestConstant.HOST_NAME, host);
-        emailReceivedServerParameters.put(EmailTestConstant.STORE_TYPE, store);
-       // emailReceivedServerParameters.put(EmailTestConstant.FOLDER_NAME, folder);
-
-        emailReceivedServerParameters.put("mail.pop3.ssl.socketFactory.class", "com.blogspot.lifeinide.AlwaysTrustSSLContextFactory");
-       // emailReceivedServerParameters.put("mail.pop3.ssl.socketFactory.port", "995");
-        emailReceivedServerParameters.put("mail.pop3.starttls.enable", "true");
-
-        return emailReceivedServerParameters;
-
-    } */
-}
