--- email/org.wso2.carbon.transport.email/src/test/java/org/wso2/carbon/transport/email/xc.java
@@ -1,7 +1,0 @@
-package org.wso2.carbon.transport.email;
-
-/**
- * Created by chathurika on 7/24/17.
- */
-public class xc {
-}
