+++ email/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>org.wso2.carbon.transport.parent</artifactId>
         <groupId>org.wso2.carbon.transport</groupId>
-        <version>4.4.13-SNAPSHOT</version>
+        <version>4.4.23-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
