+++ features/email/org.wso2.carbon.transport.email.feature/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>org.wso2.carbon.transport.email.feature.parent</artifactId>
         <groupId>org.wso2.carbon.transport</groupId>
-        <version>4.4.13-SNAPSHOT</version>
+        <version>4.4.23-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
