+++ features/email/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>org.wso2.carbon.transport.feature.parent</artifactId>
         <groupId>org.wso2.carbon.transport</groupId>
-        <version>4.4.13-SNAPSHOT</version>
+        <version>4.4.23-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
